move tests to the right place
